+++ sipx/sdp/__init__.py
@@ -1,0 +1,18 @@
+from sipx.sdp.model import AudioMedia, SdpCodec, SdpDirection, SessionDescription
+from sipx.sdp.offer_answer import (
+    SdpNegotiationError,
+    create_audio_offer,
+    create_audio_answer,
+)
+from sipx.sdp.parser import parse_sdp
+
+__all__ = [
+    "AudioMedia",
+    "SdpCodec",
+    "SdpDirection",
+    "SdpNegotiationError",
+    "SessionDescription",
+    "create_audio_answer",
+    "create_audio_offer",
+    "parse_sdp",
+]